requirements.txt
minio==7.2.5
fastapi==0.110.0
pandas==2.2.3
pyarrow==19.0.0
pyspark==3.5.1
python-dotenv==1.0.1
requests==2.32.3
s3fs==2025.3.2
uvicorn[standard]==0.29.0
